+++ changelog.md
@@ -7,5 +7,6 @@
 ### 0.0.2 - Player orientation
 
 ### 0.0.1 - Player movement
+- 5b49cea - Move entities up/down
 - 4d0aaae - cap FPS
-- b6e5814 - Translate player left/right in 2d space 
+- b6e5814 - Translate entities left/right in 2d space